Print output instead of printing "True"
This will make debugging much easier.

I also want to test that the resulting expression has the proper type,
but for this I have to duplicate the input, because of staging
restrictions. I also have to comment out the test for now, because type
errors show up as compilation errors instead of test failures.