fix the parser
I knew the using type synonyms instead of newtypes for In and Out would
lead to issues like this one, in which I swapped an In with an Out.

With only one such issue encountered so far, am still not convinced that
the extra safety would be worth the extra verbosity caused by the
wrapping and unwrapping calls.